+++ workspaces/scaffolder-backend-module-sonarqube/plugins/scaffolder-backend-module-sonarqube/src/index.ts
@@ -15,7 +15,7 @@
  */
 
 /**
- * The sonarqube module for @backstage/plugin-scaffolder-backend.
+ * The sonarqube module for \@backstage/plugin-scaffolder-backend.
  *
  * @packageDocumentation
  */