+++ workspaces/rbac/plugins/rbac-node/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## @backstage-community/plugin-rbac-node [1.4.0](https://github.com/janus-idp/backstage-plugins/compare/@backstage-community/plugin-rbac-node@1.3.1...@backstage-community/plugin-rbac-node@1.4.0) (2024-07-26)
 
+## 1.8.3
+
+### Patch Changes
+
+- 4b3653a: Clean up api report warnings and remove unnecessary files
+
 ## 1.8.2
 
 ### Patch Changes